go: simplify int types

+++ _examples/animation/animation.go
@@ -3,7 +3,7 @@ package main
 import "github.com/life4/firefly-go/firefly"
 
 var pos = firefly.Point{X: 60, Y: 40}
-var dir int32 = 2
+var dir int = 2
 
 const diameter = 20
 
@@ -14,7 +14,7 @@ func init() {
 
 func update() {
 	size := firefly.GetScreenSize()
-	if pos.X <= 0 || pos.X+diameter >= int32(size.W) {
+	if pos.X <= 0 || pos.X+diameter >= size.W {
 		dir = -dir
 	}
 	pos.X += dir
