+++ _examples/touchpad/touchpad.go
@@ -15,13 +15,13 @@ func init() {
 func update() {
 	screen := firefly.GetScreenSize()
 	center = firefly.Point{
-		X: int32(screen.W / 2),
-		Y: int32(screen.H / 2),
+		X: screen.W / 2,
+		Y: screen.H / 2,
 	}
 	input := firefly.ReadLeft()
 	point = firefly.Point{
-		X: center.X + int32(input.X/20) - radius,
-		Y: center.Y - int32(input.Y/20) - radius,
+		X: center.X + input.X/20 - radius,
+		Y: center.Y - input.Y/20 - radius,
 	}
 }
